[MacOS] Move NativeListViewBase to shared code
The main layouting logic in VirtualizingPanelLayout will be made MacOS-compatible in a subsequent commit. ListViewBaseSource is not correctly supported yet.
