+++ src/Uno.UI/UI/Xaml/Controls/ItemsWrapGrid/ItemsWrapGrid.cs
@@ -1,4 +1,4 @@
-﻿#if !NET461 && !__WASM__ && !__MACOS__
+﻿#if !NET461 && !__WASM__
 using System;
 using System.Collections.Generic;
 using System.Text;
