+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/VirtualizingPanelLayout.iOSmacOS.cs
